Unified pitch layout between FET and the game
