+++ apps/client/package.json
@@ -24,6 +24,7 @@
         "react": "^19.0.0",
         "react-dom": "^19.0.0",
         "@game/server": "workspace:*",
+        "@game/pitch-ui": "workspace:*",
         "uuid": "^11.1.0",
         "zod": "^3.24.2"
     },
