+++ apps/client/vite.config.ts
@@ -1,5 +1,6 @@
 import react from "@vitejs/plugin-react-swc"
 import { defineConfig, loadEnv } from "vite"
+import { resolve } from "path"
 
 const BASE_PATH = "/game/"
 
@@ -16,5 +17,10 @@ export default defineConfig(({ command, mode }) =>
         server: {
             allowedHosts: allowedHost ? [allowedHost] : undefined,
         },
+        resolve: {
+            alias: {
+                "@game/pitch-ui": resolve(__dirname, "../../packages/pitch-ui/src/index.ts")
+            }
+        }
     }
 })
