+++ apps/fet-editor/package.json
@@ -13,6 +13,7 @@
         "test": "vitest run --coverage"
     },
     "dependencies": {
+        "@game/pitch-ui": "workspace:*",
         "react": "^19.0.0",
         "react-dom": "^19.0.0"
     },
