+++ packages/pitch-ui/tsconfig.json
@@ -1,0 +1,14 @@
+{
+  "extends": "tsconfig/react.json",
+  "compilerOptions": {
+    "baseUrl": ".",
+    "paths": {}
+  },
+  "include": [
+    "src"
+  ],
+  "exclude": [
+    "dist",
+    "node_modules"
+  ]
+}
